requirements.txt
aniso8601==9.0.1
aws-shell==0.2.2
awscli==1.32.104
bandit==1.7.5
black==23.3.0
blinker==1.8.2
boto3==1.34.104
botocore==1.34.104
click==8.1.3
colorama==0.4.6
configobj==5.0.8
distlib==0.3.8
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
exceptiongroup==1.1.1
filelock==3.15.4
flake8==4.0.1
Flask==2.1.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.3
importlib-metadata==4.11.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdurl==0.1.2
mongoengine==0.24.1
mongomock==4.1.2
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
packaging==23.1
pathspec==0.11.1
pbr==5.11.1
platformdirs==4.2.2
pluggy==1.0.0
prompt-toolkit==1.0.18
pyasn1==0.6.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.15.1
pymongo==4.1.1
pytest==7.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==6.0.1
rich==13.3.4
rsa==4.7.2
s3transfer==0.10.1
sentinels==1.0.0
six==1.16.0
smmap==5.0.0
stevedore==5.0.0
tomli==2.0.1
urllib3==2.2.1
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
